requirements.txt

imageio==2.5.0
lz4==3.1.3
numpy==1.17.2
opencv-python==3.4.1.15
Pillow==8.2.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.2.1
tensorboardX==2.4
torch==1.7.0
torchvision==0.8.0
tqdm==4.31.1

